lead_field_eeg_fem: L_eeg = -T' * G, not T' * G.
In the prior code where the interpolation (matrix G) was built into the PCG
iteration, the column-by-column construction logic was as follows:

	L_eeg = - T(:,column index)' * G ,

whereas in the new routine the effects of the interpolation were applied with
a single matrix product

	L_eeg = T' * G;

at the end of the file. Notice the lack of the minus sign.