zef_transfer_matrix: takes a Schur expression as its last parameter
This 2-place function handle `schur_expression(Tcol,ind)` specifies how the
columns of the Schur complement of the transfer matrix `T` are to be
calculated, based on the columns of `T` or `Tcol`. This means that the
function handle needs to be defined at the call site of `zef_transfer_matrix`,
so that `B` and `C` might be captured in the closure. This definition is now
done in the files `lead_field_{eeg,tes}_fem.m`.