zeffiro_interface: zef file paths to char vectors, but also checkout files that use these from main_development_branch
It turns out the paths are further manipulated in the checked out files in a
way that is not compatible with proper strings. It is better to manipulate the
paths as strings in zeffiro_interface and then convert them to char vectors
before calling legacy code, as many changes would be required otherwise.